+++ apps/storybook/src/stories/package/wallet-connector-privy/wallet-connector-privy.stories.tsx
@@ -1,15 +1,9 @@
 import { useState } from "react";
 import { StoryObj } from "@storybook/react-vite";
 import { TradingPage } from "@orderly.network/trading";
-import {
-  Box,
-  ExtensionPositionEnum,
-  Flex,
-  installExtension,
-  Text,
-} from "@orderly.network/ui";
 import { MainNavWidget } from "@orderly.network/ui-scaffold";
 import { Scaffold } from "@orderly.network/ui-scaffold";
+import { useNav } from "../../../hooks/useNav";
 import { useOrderlyConfig } from "../../../hooks/useOrderlyConfig";
 import { OrderlyIcon } from "../trading/icons.tsx";
 
@@ -17,10 +11,14 @@ const meta = {
   title: "Package/wallet-connector-privy",
   component: TradingPage,
   decorators: [
-    (Story) => {
+    (Story: any) => {
       const config = useOrderlyConfig();
+      const { onRouteChange } = useNav();
       return (
-        <Scaffold mainNavProps={config.scaffold.mainNavProps}>
+        <Scaffold
+          mainNavProps={config.scaffold.mainNavProps}
+          routerAdapter={{ onRouteChange }}
+        >
           <Story />
         </Scaffold>
       );
@@ -50,6 +48,7 @@ const meta = {
 };
 
 export default meta;
+
 type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {