+++ packages/markets/src/components/fundingComparison/fundingComparison.mobile.ui.tsx
@@ -17,10 +17,12 @@ export const MobileFundingComparison: FC<FundingComparisonReturn> = (props) => {
           className: cn("oui-h-[34px] oui-cursor-pointer"),
         };
       }}
+      onSort={props.onSort}
       classNames={{
         header: "oui-h-9",
         body: "oui-text-2xs",
       }}
+      manualSorting
     />
   );
 };