+++ packages/markets/src/components/marketsProvider/index.tsx
@@ -5,6 +5,7 @@ import {
   useState,
   useContext,
   useCallback,
+  useMemo,
 } from "react";
 import { API } from "@orderly.network/types";
 import { LeftNavProps, RouterAdapter } from "@orderly.network/ui-scaffold";
@@ -30,28 +31,44 @@ export type MarketsProviderProps = {
     routerAdapter?: RouterAdapter;
     leftNav?: LeftNavProps;
   };
+  comparisonProps?: {
+    /**
+     * Set Name of Exchanges in the comparison list.
+     * @default 'Orderly'
+     */
+    exchangesName?: string;
+    /**
+     * Set Icon URL of Exchanges in the comparison list.
+     * @default ""
+     */
+    exchangesIconSrc?: string;
+  };
 };
 
 export const MarketsProvider: FC<PropsWithChildren<MarketsProviderProps>> = (
   props,
 ) => {
+  const { symbol, comparisonProps, children, onSymbolChange } = props;
   const [searchValue, setSearchValue] = useState("");
 
   const clearSearchValue = useCallback(() => {
     setSearchValue("");
   }, []);
 
+  const memoizedValue = useMemo<MarketsContextState>(() => {
+    return {
+      searchValue,
+      onSearchValueChange: setSearchValue,
+      clearSearchValue,
+      symbol: symbol,
+      onSymbolChange: onSymbolChange,
+      comparisonProps: comparisonProps,
+    };
+  }, [searchValue, symbol, onSymbolChange, setSearchValue, comparisonProps]);
+
   return (
-    <MarketsContext.Provider
-      value={{
-        searchValue,
-        onSearchValueChange: setSearchValue,
-        clearSearchValue,
-        symbol: props.symbol,
-        onSymbolChange: props.onSymbolChange,
-      }}
-    >
-      {props.children}
+    <MarketsContext.Provider value={memoizedValue}>
+      {children}
     </MarketsContext.Provider>
   );
 };