+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @orderly/docs
 
+## 1.5.2-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/hooks@2.5.2-alpha.0
+  - @orderly.network/web3-onboard@2.5.2-alpha.0
+  - @orderly.network/types@2.5.2-alpha.0
+
+## 1.5.1
+
+### Patch Changes
+
+- publish
+- 5ad8a1b: publish
+- Updated dependencies
+- Updated dependencies [5ad8a1b]
+  - @orderly.network/hooks@2.5.1
+  - @orderly.network/web3-onboard@2.5.1
+  - @orderly.network/types@2.5.1
+
+## 1.5.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/hooks@2.5.1-alpha.0
+  - @orderly.network/web3-onboard@2.5.1-alpha.0
+  - @orderly.network/types@2.5.1-alpha.0
+
 ## 1.5.0
 
 ### Minor Changes