+++ packages/cli-codemod/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @orderly.network/codemod
 
+## 1.5.2-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 1.5.1
+
+### Patch Changes
+
+- publish
+- 5ad8a1b: publish
+
+## 1.5.1-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 1.5.0
 
 ### Minor Changes