+++ packages/core/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @orderly.network/core
 
+## 2.5.2-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/types@2.5.2-alpha.0
+
+## 2.5.1
+
+### Patch Changes
+
+- publish
+- 5ad8a1b: publish
+- Updated dependencies
+- Updated dependencies [5ad8a1b]
+  - @orderly.network/types@2.5.1
+
+## 2.5.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/types@2.5.1-alpha.0
+
 ## 2.5.0
 
 ### Minor Changes