+++ packages/create-orderly-app/CHANGELOG.md
@@ -1,5 +1,24 @@
 # create-orderly-app
 
+## 1.10.2-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 1.10.1
+
+### Patch Changes
+
+- publish
+- 5ad8a1b: publish
+
+## 1.10.1-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 1.10.0
 
 ### Minor Changes