+++ packages/eslint-config-custom/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @orderly.network/eslint-config
 
+## 2.5.2-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 2.5.1
+
+### Patch Changes
+
+- publish
+- 5ad8a1b: publish
+
+## 2.5.1-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 2.5.0
 
 ### Minor Changes