+++ packages/perp/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @orderly.network/futures
 
+## 4.5.2-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/types@2.5.2-alpha.0
+  - @orderly.network/utils@2.5.2-alpha.0
+
+## 4.5.1
+
+### Patch Changes
+
+- publish
+- 5ad8a1b: publish
+- Updated dependencies
+- Updated dependencies [5ad8a1b]
+  - @orderly.network/types@2.5.1
+  - @orderly.network/utils@2.5.1
+
+## 4.5.1-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/types@2.5.1-alpha.0
+  - @orderly.network/utils@2.5.1-alpha.0
+
 ## 4.5.0
 
 ### Minor Changes